requirements.txt
Django==1.7.5
PyMySQL==0.6.6
SQLAlchemy==0.9.8
Werkzeug==0.10.1
autopep8==1.1.1
beautifulsoup4==4.3.2
defusedxml==0.4.1
django-admin-tools==0.5.2
django-debug-toolbar==1.2.2
django-extensions==1.5.0
djangorestframework==3.1.0
djangorestframework-xml==1.0.1
docutils==0.12
feedparser==5.1.3
lxml==3.4.2
pep8==1.6.2
selenium==2.45.0
simplejson==3.6.5
six==1.9.0
sqlparse==0.1.14
sqlsoup==0.9.0
wsgiref==0.1.2
